% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2hex.R
\name{col2hex}
\alias{col2hex}
\title{A function to convert quoted names of colors in R to their Hex values}
\usage{
col2hex(color)
}
\arguments{
\item{color}{The quoted name of the color whose hex code you want}
}
\description{
A function to convert quoted names of colors in R to their Hex values
}
\examples{
col2hex("wheat")

}
