% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_water_year_type.R
\name{get_water_year_type}
\alias{get_water_year_type}
\title{Water year classifications}
\usage{
get_water_year_type()
}
\description{
Get water year classifications from CDEC webpage
}
\examples{
df <- get_water_year_type()
}
