% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/len.R
\name{len}
\alias{len}
\title{An abbreviated form of length(unique())}
\usage{
len(x)
}
\arguments{
\item{x}{The object you want to find the number of unique things of}
}
\description{
An abbreviated form of length(unique())
}
\examples{
len(letters)

}
