% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timer}
\alias{timer}
\title{A simple timer function.}
\usage{
timer(interval = 1, units = "secs", sound = 2)
}
\arguments{
\item{interval}{The length of time you want}

\item{units}{The unit of time you want to use, i.e. "secs", "mins", "hours"}

\item{sound}{Your sound of choice from beepr::beep}
}
\value{
a sound from the beepr pkg; options are 1-11.  Defaults to 2
}
\description{
A simple timer function.
}
\examples{
timer(45, "mins")

}
