% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_year.R
\name{water_year}
\alias{water_year}
\title{Get water year based on numeric year and numeric month}
\usage{
water_year(y, m)
}
\arguments{
\item{y}{Numeric year}

\item{m}{Numeric month (i.e., 1 = Jan, 2 = Feb, etc.)}
}
\description{
Get water year based on numeric year and numeric month
}
\examples{
water_year(2016, 9)
water_year(2016, 10)

}
