% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_year_type}
\alias{water_year_type}
\title{Water year type}
\format{
A data frame with 4 columns:
\describe{
  \item{WaterYear}{1901-present}
  \item{Valley}{SAC or SJR}
  \item{WYT}{Water year classifications as abbreviations: C, D, BN, AN, W}
  \item{WaterYearType}{Water year classifications: Critical, Dry, Below Normal, Above Normal, Wet.}
}
}
\usage{
water_year_type
}
\description{
Water year classifications for the Sacramento and San Joaquin valleys from http://cdec.water.ca.gov/reportapp/javareports?name=WSIHIST.
}
\keyword{datasets}
