% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wy_date.R
\name{wy_date}
\alias{wy_date}
\title{Water year date}
\usage{
wy_date(x, y)
}
\arguments{
\item{x}{Day of water year}

\item{y}{Water year}
}
\description{
Get date for day of water year
}
\examples{
wy_date(x = seq(1, 361, 60), y = 2011)

}
