% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wy_week.R
\name{wy_week}
\alias{wy_week}
\title{Water year week}
\usage{
wy_week(x)
}
\arguments{
\item{x}{A date-time object.}
}
\description{
Get week in water year from a date-time object.
}
\details{
Returns the number of complete seven day periods that have occurred between the date and October 1st, plus one. Based on lubridate::week().
}
\examples{
wy_week(as.Date(c("2016-10-01", "2017-01-01")))

}
