% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wy_yday.R
\name{wy_yday}
\alias{wy_yday}
\title{Day of water year}
\usage{
wy_yday(x)
}
\arguments{
\item{x}{A date-time object.}
}
\description{
Get day of water year from date-time object.
Warning: uses the system-specific time zone and not the time zone associated with the date-time object.
}
\examples{
wy_yday(as.Date(c("2016-01-01", "2016-10-01", "2017-01-01")))
wy_yday(as.Date(c("2016-03-01", "2017-03-01"))) # 2016 was a leap year

}
