% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_water_year_type.R
\name{get_water_year_type}
\alias{get_water_year_type}
\title{Water year classifications}
\usage{
get_water_year_type(water_year, valley = c("SAC", "SJR"))
}
\arguments{
\item{water_year}{numeric vector of water years}

\item{valley}{Sacramento or San Joaquin valley (options: SAC or SJR)}
}
\description{
Get water year classifications
}
\examples{
get_water_year_type(1900:1906)
get_water_year_type(1900:1906, "SJR")
get_water_year_type(2014:2020)
}
