% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timer}
\alias{timer}
\title{A simple timer function.}
\usage{
timer(interval, units)
}
\arguments{
\item{interval}{The length of time you want}

\item{units}{The unit of time you want to use, i.e. "secs", "mins", "hours"}
}
\description{
A simple timer function.
}
\examples{
timer(45, "mins")

}
