% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_year.R
\name{water_year}
\alias{water_year}
\title{Water year}
\usage{
water_year(x)
}
\arguments{
\item{x}{A date-time object.}
}
\description{
Get water year from date-time object.
Warning: uses the system-specific time zone and not the time zone associated with the date-time object.
}
\examples{
x <- c("2016-12-31", "2017-01-01", "2017-09-30", "2017-10-01")
water_year(as.Date(x))
water_year(as.Date(x)) == format(as.Date(x), "\%Y")
}
